+++ recipes/configure.rb
@@ -1,4 +1,12 @@
-# Overide the solr.in.sh script we get from the install script
+include_recipe 'clocker'
+
+# Take a cluster wide lock on restarting the solr node
+clocker 'solr-node-restart' do
+  lockid 'solr-node-restart'
+  lockwait 30
+  zookeeper node['jlsolrcloud']['zkhosts'].sample
+  action :clockon
+end
 
 if node['jlsolrcloud']['zkhosts'].length < 1
   fail "At least one zkhost is needed in node['jlsolrcloud']['zkhosts']"
@@ -23,6 +31,7 @@ if node['jlsolrcloud']['solr_home_override']
     user   node['jlsolrcloud']['user']
     group  node['jlsolrcloud']['group']
     mode   0755
+    only_if { Clocker.held?('solr-node-restart', run_context) }
     notifies :restart, 'service[solr]'
   end
 else
@@ -31,6 +40,7 @@ else
     user   node['jlsolrcloud']['user']
     group  node['jlsolrcloud']['group']
     mode   0755
+    only_if { Clocker.held?('solr-node-restart', run_context) }
     notifies :restart, 'service[solr]'
   end
 
@@ -45,6 +55,7 @@ template "#{node['jlsolrcloud']['solr_home']}/solr.in.sh" do
   user  node['jlsolrcloud']['user']
   group node['jlsolrcloud']['group']
   mode  0755
+  only_if { Clocker.held?('solr-node-restart', run_context) }
   notifies :restart, 'service[solr]'
 end
 
@@ -52,9 +63,12 @@ cookbook_file '/var/solr/log4j.properties' do
   user  node['jlsolrcloud']['user']
   group node['jlsolrcloud']['group']
   mode  0755
+  only_if { Clocker.held?('solr-node-restart', run_context) }
   notifies :restart, 'service[solr]'
 end
 
+# We do a single line change to the solr script to allow us to rotate
+# the console log using copy/truncate methods without restarting solr
 cookbook_file '/opt/solr/bin/solr' do
   user  node['jlsolrcloud']['user']
   group node['jlsolrcloud']['group']
@@ -62,7 +76,15 @@ cookbook_file '/opt/solr/bin/solr' do
   # notify immediatly so that on first run Solr will
   # be running as directed for any other chef things
   # that may want to use it's API
+  only_if { Clocker.held?('solr-node-restart', run_context) }
   notifies :restart, 'service[solr]', :immediately
+end
+
+# Release the solr node restart lock
+clocker 'solr-node-restart' do
+  lockid 'solr-node-restart'
+  zookeeper node['jlsolrcloud']['zkhosts'].sample
+  action :clockoff
 end
 
 service 'solr' do