+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.6.0
+ - Use clocker to only restart one node on the cluster at a time
+ - Correct log4j log location
+
 ### 0.4.0
  - Add attributes to control logrotation of console logs, change defaults to 
    weekly and keep 3 log files